Define separate ecto upload schema
